refactor AddressFactory out into its own file/class. (shush the spec)
