+++ spec/db/address_factory_spec.rb
@@ -1,6 +1,7 @@
 require 'rails_helper'
 
-require_relative '../../db/seed_helpers'
+require_relative File.join(Rails.root, 'db/seed_helpers/address_factory')
+
 
 ENV_SEED_FAKE_CSV_FNAME_KEY = 'SHF_SEED_FAKE_ADDR_CSV_FILE' unless defined?(ENV_SEED_FAKE_CSV_FNAME_KEY)
 
@@ -22,7 +23,7 @@ RSpec.describe SeedHelper::AddressFactory do
 
   before(:all) do
     DatabaseCleaner.start
-    
+
     create_empty_file(SEED_DB_DIR, EMPTY_CSV_FILENAME)
     create_csv_file(SEED_DB_DIR, AF_FAKE_ADDRESSES_CSV_FILENAME, AF_FAKE_ADDRESSES)
   end
@@ -66,7 +67,7 @@ RSpec.describe SeedHelper::AddressFactory do
           env.delete(ENV_SEED_FAKE_CSV_FNAME_KEY)
           stub_const('ENV', env)
 
-          expect(address_factory.fake_addresses_csv_filename).to eq File.join(Rails.root, 'db', SeedHelper::DEFAULT_FAKE_ADDR_FILENAME)
+          expect(address_factory.fake_addresses_csv_filename).to eq File.join(Rails.root, 'db', described_class::DEFAULT_FAKE_ADDR_FILENAME)
         end
 
         ENV[ENV_SEED_FAKE_CSV_FNAME_KEY] = orig_env_csv_fn unless orig_env_csv_fn.nil?
@@ -78,6 +79,7 @@ RSpec.describe SeedHelper::AddressFactory do
       it 'gets the CSV filename from the ENV value and adds the db directory' do
         RSpec::Mocks.with_temporary_scope do
           stub_const('ENV', ENV.to_hash.merge({ ENV_SEED_FAKE_CSV_FNAME_KEY => 'fake_addresses.csv' }))
+          allow_any_instance_of(SeedHelper::AddressFactory).to receive(:tell).and_return(false)
 
           expect(address_factory.fake_addresses_csv_filename).to eq File.join(Rails.root, 'db', 'fake_addresses.csv')
         end
@@ -92,6 +94,7 @@ RSpec.describe SeedHelper::AddressFactory do
 
       RSpec::Mocks.with_temporary_scope do
         stub_const('ENV', ENV.to_hash.merge({ ENV_SEED_FAKE_CSV_FNAME_KEY => AF_FAKE_ADDRESSES_CSV_FILENAME }))
+        allow_any_instance_of(SeedHelper::AddressFactory).to receive(:tell).and_return(false)
 
         expect(address_factory.already_constructed_addresses.size).to eq 4
         classes_read = address_factory.already_constructed_addresses.map(&:class).uniq
@@ -129,6 +132,7 @@ RSpec.describe SeedHelper::AddressFactory do
 
       it 'if there are no already constructed addresses, an Address is created and geocoded' do
         allow(address_factory).to receive(:already_constructed_addresses).and_return([])
+        allow_any_instance_of(SeedHelper::AddressFactory).to receive(:tell).and_return(false)
 
         expect(address_factory).to receive(:create_a_new_address).and_call_original
         expect(Geocoder).to receive(:search).at_least(1).times
